requirements.txt
setuptools>=41.0.0
tensorflow
# if your jaxlib is different, modify this two.
chex==0.1.5
flax==0.6.0
dm_env>=1.6
matplotlib>=3.7.1
dm-tree>=0.1.8
immutabledict>=2.2.3
Pillow>=9.4.0
mediapy
tqdm>=4.65.0
absl-py>=1.4.0
lightning==2.3.3
pytorch-lightning==2.3.3
hydra-core==1.3.2
shapely==2.0.5
tensorboard
Cython==3.0.11
gymnasium==0.29.1
prettytable==3.10.2
einops==0.8.0
transformers==4.42.4
gym==0.26.2
typing_extensions>=4.6.0